requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3
pyTelegramBotAPI==4.14.0
requests==2.31.0